memory_ai_agent/requirements.txt
requests>=2.28.1
rich>=13.3.5
duckduckgo-search>=3.8.3
python-dotenv>=1.0.0